Move package handler to handlers
