+++ gengokit/handlers/handler_test.go
@@ -1,4 +1,4 @@
-package handler
+package handlers
 
 import (
 	"go/ast"
