+++ gengokit/handlers/hooks.go
@@ -1,4 +1,4 @@
-package handler
+package handlers
 
 import (
 	"io"
