+++ gengokit/handlers/middlewares_test.go
@@ -1,4 +1,4 @@
-package handler
+package handlers
 
 import (
 	"io/ioutil"
