+++ gengokit/handlers/templates/endpoint_middlewares.go
